Add Java support for `selectable`, `offset`, and `closeAreals` properties on `VectorInfo` and `WideVectorInfo`. Add matching getters for setters to enable Kotlin property syntax. Cleanup and refactor JNI code.

+++ android/library/maply/src/main/java/com/mousebird/maply/BaseInfo.java
@@ -93,7 +93,7 @@ public class BaseInfo
     /**
      * @return The Z offset for geometry.  Rarely used anymore.
      */
-    public native double getDrawOffset();
+    public native float getDrawOffset();
 
     /**
      * Set the drawPriority for the geometry.  Draw priority controls the order
